+++ src/finiteVolume/cellCentred/operators/gaussGreenLaplacian.cpp
@@ -6,6 +6,7 @@
 
 #include "NeoN/core/parallelAlgorithms.hpp"
 #include "NeoN/finiteVolume/cellCentred/operators/gaussGreenLaplacian.hpp"
+#include "NeoN/linearAlgebra/meshIterationStrategies.hpp"
 
 namespace NeoN::finiteVolume::cellCentred
 {
